refactor: modified directory name.
