+++ spring-cloud-alibaba-starters/spring-cloud-starter-alibaba-nacos-discovery/src/main/resources/META-INF/spring.factories
@@ -7,7 +7,7 @@ org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
   com.alibaba.cloud.nacos.discovery.configclient.NacosConfigServerAutoConfiguration,\
   com.alibaba.cloud.nacos.loadbalancer.LoadBalancerNacosAutoConfiguration,\
   com.alibaba.cloud.nacos.NacosServiceAutoConfiguration,\
-  com.alibaba.cloud.nacos.utils.UtilIPv6AutoConfiguration
+  com.alibaba.cloud.nacos.util.UtilIPv6AutoConfiguration
 org.springframework.cloud.bootstrap.BootstrapConfiguration=\
   com.alibaba.cloud.nacos.discovery.configclient.NacosDiscoveryClientConfigServiceBootstrapConfiguration
 org.springframework.boot.SpringApplicationRunListener=\
